requirements.txt
tensorflow 
pydot
tensorflow_model_optimization
pocketsphinx
gtts
SpeechRecognition
pyaudio
openai-whisper